pull and merge from official master branch

+++ manager/src/main/java/org/openremote/manager/rules/RulesResourceImpl.java
@@ -26,15 +26,16 @@ import org.openremote.manager.web.ManagerWebResource;
 import org.openremote.model.Constants;
 import org.openremote.model.asset.Asset;
 import org.openremote.model.query.AssetQuery;
-import org.openremote.model.asset.UserAsset;
+import org.openremote.model.asset.UserAssetLink;
 import org.openremote.model.http.RequestParams;
 import org.openremote.model.query.RulesetQuery;
 import org.openremote.model.rules.*;
 import org.openremote.model.rules.geofence.GeofenceDefinition;
 import org.openremote.model.security.ClientRole;
-import org.openremote.model.security.Tenant;
+import org.openremote.model.security.Realm;
 
 import javax.ws.rs.BeanParam;
+import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Response;
 import java.util.List;
@@ -74,12 +75,12 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
     }
 
     @Override
-    public RulesEngineInfo getTenantEngineInfo(RequestParams requestParams, String realm) {
+    public RulesEngineInfo getRealmEngineInfo(RequestParams requestParams, String realm) {
         if (!isRealmAccessibleByUser(realm) || isRestrictedUser()) {
             throw new WebApplicationException(Response.Status.FORBIDDEN);
         }
 
-        RulesEngine<TenantRuleset> engine = rulesService.tenantEngines.get(realm);
+        RulesEngine<RealmRuleset> engine = rulesService.realmEngines.get(realm);
         return getEngineInfo(engine);
     }
 
@@ -139,7 +140,7 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
     }
 
     @Override
-    public TenantRuleset[] getTenantRulesets(@BeanParam RequestParams requestParams, String realm, List<Ruleset.Lang> languages, boolean fullyPopulate) {
+    public RealmRuleset[] getRealmRulesets(@BeanParam RequestParams requestParams, String realm, List<Ruleset.Lang> languages, boolean fullyPopulate) {
 
         if (isAuthenticated() && !isRealmAccessibleByUser(realm)) {
             throw new WebApplicationException(Response.Status.FORBIDDEN);
@@ -147,8 +148,8 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
 
         boolean publicOnly = !isAuthenticated() || isRestrictedUser() | !hasResourceRole(ClientRole.READ_RULES.getValue(), Constants.KEYCLOAK_CLIENT_ID);
 
-        List<TenantRuleset> result = rulesetStorageService.findAll(
-            TenantRuleset.class,
+        List<RealmRuleset> result = rulesetStorageService.findAll(
+            RealmRuleset.class,
             new RulesetQuery().
                 setRealm(realm)
                 .setLanguages(languages.toArray(new Ruleset.Lang[0]))
@@ -165,7 +166,7 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
                 })
         );
 
-        return result.toArray(new TenantRuleset[0]);
+        return result.toArray(new RealmRuleset[0]);
     }
 
     @Override
@@ -246,48 +247,54 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
     /* ################################################################################################# */
 
     @Override
-    public long createTenantRuleset(@BeanParam RequestParams requestParams, TenantRuleset ruleset) {
-        Tenant tenant = identityService.getIdentityProvider().getTenant(ruleset.getRealm());
-        if (tenant == null) {
+    public long createRealmRuleset(@BeanParam RequestParams requestParams, RealmRuleset ruleset) {
+        Realm realm = identityService.getIdentityProvider().getRealm(ruleset.getRealm());
+        if (realm == null) {
             throw new WebApplicationException(BAD_REQUEST);
         }
-        if (!isTenantActiveAndAccessible(tenant) || isRestrictedUser()) {
-            LOG.info("Forbidden access for user '" + getUsername() + "': " + tenant);
+        if (!isRealmActiveAndAccessible(realm) || isRestrictedUser()) {
+            LOG.fine("Realm '" + realm + "' is nonexistent, inactive or inaccessible: username=" + getUsername());
             throw new WebApplicationException(Response.Status.FORBIDDEN);
+        }
+
+        // Only super users can create/modify groovy rules
+        // TODO: Implement robust groovy sandbox
+        if (ruleset.getLang() == Ruleset.Lang.GROOVY && !isSuperUser()) {
+            throw new ForbiddenException("Only super users can create/modify groovy rules for security reasons");
         }
         ruleset = rulesetStorageService.merge(ruleset);
         return ruleset.getId();
     }
 
     @Override
-    public TenantRuleset getTenantRuleset(@BeanParam RequestParams requestParams, Long id) {
-        TenantRuleset ruleset = rulesetStorageService.find(TenantRuleset.class, id);
+    public RealmRuleset getRealmRuleset(@BeanParam RequestParams requestParams, Long id) {
+        RealmRuleset ruleset = rulesetStorageService.find(RealmRuleset.class, id);
         if (ruleset == null) {
             throw new WebApplicationException(NOT_FOUND);
         }
-        Tenant tenant = identityService.getIdentityProvider().getTenant(ruleset.getRealm());
-        if (tenant == null) {
+        Realm realm = identityService.getIdentityProvider().getRealm(ruleset.getRealm());
+        if (realm == null) {
             throw new WebApplicationException(BAD_REQUEST);
         }
-        if (!isTenantActiveAndAccessible(tenant) || isRestrictedUser()) {
-            LOG.fine("Forbidden access for user '" + getUsername() + "': " + tenant);
+        if (!isRealmActiveAndAccessible(realm) || isRestrictedUser()) {
+            LOG.fine("Forbidden access for user '" + getUsername() + "': " + realm);
             throw new WebApplicationException(Response.Status.FORBIDDEN);
         }
         return ruleset;
     }
 
     @Override
-    public void updateTenantRuleset(@BeanParam RequestParams requestParams, Long id, TenantRuleset ruleset) {
-        TenantRuleset existingRuleset = rulesetStorageService.find(TenantRuleset.class, id);
+    public void updateRealmRuleset(@BeanParam RequestParams requestParams, Long id, RealmRuleset ruleset) {
+        RealmRuleset existingRuleset = rulesetStorageService.find(RealmRuleset.class, id);
         if (existingRuleset == null) {
             throw new WebApplicationException(NOT_FOUND);
         }
-        Tenant tenant = identityService.getIdentityProvider().getTenant(existingRuleset.getRealm());
-        if (tenant == null) {
+        Realm realm = identityService.getIdentityProvider().getRealm(existingRuleset.getRealm());
+        if (realm == null) {
             throw new WebApplicationException(BAD_REQUEST);
         }
-        if (!isTenantActiveAndAccessible(tenant) || isRestrictedUser()) {
-            LOG.fine("Forbidden access for user '" + getUsername() + "': " + tenant);
+        if (!isRealmActiveAndAccessible(realm) || isRestrictedUser()) {
+            LOG.fine("Forbidden access for user '" + getUsername() + "': " + realm);
             throw new WebApplicationException(Response.Status.FORBIDDEN);
         }
         if (!id.equals(ruleset.getId())) {
@@ -296,23 +303,36 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
         if (!existingRuleset.getRealm().equals(ruleset.getRealm())) {
             throw new WebApplicationException("Requested realm and existing ruleset realm must match", BAD_REQUEST);
         }
+
+        // Only super users can create/modify groovy rules
+        // TODO: Implement robust groovy sandbox
+        if (ruleset.getLang() == Ruleset.Lang.GROOVY && !isSuperUser()) {
+            throw new ForbiddenException("Only super users can create/modify groovy rules for security reasons");
+        }
         rulesetStorageService.merge(ruleset);
     }
 
     @Override
-    public void deleteTenantRuleset(@BeanParam RequestParams requestParams, Long id) {
-        TenantRuleset ruleset = rulesetStorageService.find(TenantRuleset.class, id);
+    public void deleteRealmRuleset(@BeanParam RequestParams requestParams, Long id) {
+        RealmRuleset ruleset = rulesetStorageService.find(RealmRuleset.class, id);
         if (ruleset == null) {
             return;
         }
-        Tenant tenant = identityService.getIdentityProvider().getTenant(ruleset.getRealm());
-        if (tenant == null) {
+        Realm realm = identityService.getIdentityProvider().getRealm(ruleset.getRealm());
+        if (realm == null) {
             throw new WebApplicationException(BAD_REQUEST);
         }
-        if (!isTenantActiveAndAccessible(tenant) || isRestrictedUser()) {
+        if (!isRealmActiveAndAccessible(realm) || isRestrictedUser()) {
             throw new WebApplicationException(Response.Status.FORBIDDEN);
         }
-        rulesetStorageService.delete(TenantRuleset.class, id);
+
+        // Only super users can create/modify groovy rules
+        // TODO: Implement robust groovy sandbox
+        if (ruleset.getLang() == Ruleset.Lang.GROOVY && !isSuperUser()) {
+            throw new ForbiddenException("Only super users can create/modify groovy rules for security reasons");
+        }
+
+        rulesetStorageService.delete(RealmRuleset.class, id);
     }
 
     /* ################################################################################################# */
@@ -333,6 +353,13 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
         if (isRestrictedUser() && !assetStorageService.isUserAsset(getUserId(), asset.getId())) {
             throw new WebApplicationException(Response.Status.FORBIDDEN);
         }
+
+        // Only super users can create/modify groovy rules
+        // TODO: Implement robust groovy sandbox
+        if (ruleset.getLang() == Ruleset.Lang.GROOVY && !isSuperUser()) {
+            throw new ForbiddenException("Only super users can create/modify groovy rules for security reasons");
+        }
+
         ruleset = rulesetStorageService.merge(ruleset);
         return ruleset.getId();
     }
@@ -379,6 +406,12 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
             throw new WebApplicationException("Can't update asset ID, delete and create the ruleset to reassign",
                                               BAD_REQUEST);
         }
+
+        // Only super users can create/modify groovy rules
+        // TODO: Implement robust groovy sandbox
+        if (ruleset.getLang() == Ruleset.Lang.GROOVY && !isSuperUser()) {
+            throw new ForbiddenException("Only super users can create/modify groovy rules for security reasons");
+        }
         rulesetStorageService.merge(ruleset);
     }
 
@@ -398,6 +431,13 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
         if (isRestrictedUser() && !assetStorageService.isUserAsset(getUserId(), asset.getId())) {
             throw new WebApplicationException(Response.Status.FORBIDDEN);
         }
+
+        // Only super users can create/modify groovy rules
+        // TODO: Implement robust groovy sandbox
+        if (ruleset.getLang() == Ruleset.Lang.GROOVY && !isSuperUser()) {
+            throw new ForbiddenException("Only super users can create/modify groovy rules for security reasons");
+        }
+
         rulesetStorageService.delete(AssetRuleset.class, id);
     }
 
@@ -407,7 +447,7 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
 
         asset = assetStorageService.find(
             new AssetQuery()
-                .select(AssetQuery.Select.selectExcludePathAndAttributes())
+                .select(new AssetQuery.Select().excludeAttributes())
                 .ids(assetId));
 
         if (asset == null)
@@ -417,7 +457,7 @@ public class RulesResourceImpl extends ManagerWebResource implements RulesResour
         if (!asset.isAccessPublicRead()) {
 
             // If asset is linked to users then only those users can get the geofences for it
-            List<UserAsset> userAssetLinks = assetStorageService.findUserAssets(asset.getRealm(), null, assetId);
+            List<UserAssetLink> userAssetLinks = assetStorageService.findUserAssetLinks(asset.getRealm(), null, assetId);
 
             if (!userAssetLinks.isEmpty()) {
                 if (!isAuthenticated() || userAssetLinks.stream().noneMatch(userAssetLink -> userAssetLink.getId().getUserId().equals(getUserId()))) {
